move jwt logic to own package
